feat(runtime): extract contracts and quickjs core into plugin-runtime package
