+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/dispatchIntent.ts
@@ -1,4 +1,4 @@
-import type { DispatchIntent } from "./quickjsContracts";
+import type { DispatchIntent } from "./contracts";
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
