+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/runtimeIdentity.ts
@@ -1,5 +1,5 @@
 import { nanoid } from "nanoid";
-import type { InstanceId, PackageId } from "./quickjsContracts";
+import type { InstanceId, PackageId } from "./contracts";
 
 export function createInstanceId(packageId: PackageId): InstanceId {
   return `${packageId}@${nanoid(8)}`;
