+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/runtimeService.integration.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, describe, expect, it } from "vitest";
-import { QuickJSRuntimeService } from "./quickjsRuntimeService";
+import { QuickJSRuntimeService } from "./runtimeService";
 
 const COUNTER_PLUGIN = `
 definePlugin(({ ui }) => {
