+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/runtimeService.ts
@@ -7,7 +7,7 @@ import type {
   LoadedPlugin,
   PackageId,
   RuntimeErrorPayload,
-} from "./quickjsContracts";
+} from "./contracts";
 import { validateUINode } from "./uiSchema";
 
 const BOOTSTRAP_SOURCE = `
