+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/uiSchema.test.ts
