+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": "."
+  },
+  "include": ["src/**/*"]
+}
