+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["client/src/**/*", "shared/**/*", "server/**/*"],
+  "include": ["client/src/**/*", "packages/plugin-runtime/src/**/*", "shared/**/*", "server/**/*"],
   "exclude": ["node_modules", "build", "dist", "**/*.test.ts"],
   "compilerOptions": {
     "incremental": true,
@@ -17,6 +17,7 @@
     "types": ["node", "vite/client"],
     "paths": {
       "@/*": ["./client/src/*"],
+      "@runtime/*": ["./packages/plugin-runtime/src/*"],
       "@shared/*": ["./shared/*"]
     }
   }
