+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/vite.config.ts
@@ -15,6 +15,7 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(import.meta.dirname, "client", "src"),
+      "@runtime": path.resolve(import.meta.dirname, "packages", "plugin-runtime", "src"),
       "@shared": path.resolve(import.meta.dirname, "shared"),
       "@assets": path.resolve(import.meta.dirname, "attached_assets"),
     },
