+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/vitest.config.ts
@@ -3,8 +3,7 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     environment: "node",
-    include: ["client/src/**/*.test.ts"],
-    exclude: ["client/src/**/*.integration.test.ts"],
+    include: ["client/src/**/*.test.ts", "packages/plugin-runtime/src/**/*.test.ts"],
+    exclude: ["client/src/**/*.integration.test.ts", "packages/plugin-runtime/src/**/*.integration.test.ts"],
   },
 });
-
