+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/vitest.integration.config.ts
@@ -3,7 +3,6 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     environment: "node",
-    include: ["client/src/**/*.integration.test.ts"],
+    include: ["client/src/**/*.integration.test.ts", "packages/plugin-runtime/src/**/*.integration.test.ts"],
   },
 });
-
