+++ ttmp/2026/02/09/WEBVM-003-DEVX-UI-PACKAGE-DOCS-OVERHAUL--developer-ui-overhaul-reusable-vm-package-and-documentation/changelog.md
@@ -66,3 +66,18 @@ Completed theme-stack unification task 3 by removing next-themes and wiring Sonn
 - /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/cmd/experiments/2026-02-08--simulated-communication/plugin-playground/pnpm-lock.yaml — Lockfile updated after dependency removal
 - /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/ttmp/2026/02/09/WEBVM-003-DEVX-UI-PACKAGE-DOCS-OVERHAUL--developer-ui-overhaul-reusable-vm-package-and-documentation/tasks.md — Marked task 3 complete
 
+
+## 2026-02-09
+
+Completed task 4 by scaffolding packages/plugin-runtime and migrating contracts, runtime identity, schema/intent validators, and QuickJS runtime service into the new package; rewired app imports and test config to consume package sources.
+
+### Related Files
+
+- /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/contracts.ts — Migrated runtime request/response and intent contracts
+- /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/runtimeIdentity.ts — Migrated instance ID generation helper
+- /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/runtimeService.ts — Migrated QuickJS runtime core
+- /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/cmd/experiments/2026-02-08--simulated-communication/plugin-playground/vite.config.ts — Added @runtime alias for package source consumption
+- /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/cmd/experiments/2026-02-08--simulated-communication/plugin-playground/vitest.config.ts — Unit tests now include package runtime tests
+- /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/cmd/experiments/2026-02-08--simulated-communication/plugin-playground/vitest.integration.config.ts — Integration tests now include package runtime tests
+- /home/manuel/workspaces/2026-02-08/plugin-playground/go-go-labs/ttmp/2026/02/09/WEBVM-003-DEVX-UI-PACKAGE-DOCS-OVERHAUL--developer-ui-overhaul-reusable-vm-package-and-documentation/tasks.md — Marked task 4 complete
+
