+++ ttmp/2026/02/09/WEBVM-003-DEVX-UI-PACKAGE-DOCS-OVERHAUL--developer-ui-overhaul-reusable-vm-package-and-documentation/tasks.md
@@ -5,7 +5,7 @@
 - [x] Remove dead/template components and unused UI wrappers/dependencies from plugin-playground
 - [x] Remove remaining debug/template leftovers (WidgetRenderer globals/logs, index analytics placeholders/comment block, unused helper imports/hooks)
 - [x] Unify theme stack (remove custom-vs-next-themes split and keep one provider model)
-- [ ] Create one `plugin-runtime` package scaffold and migrate contracts + QuickJS core into it
+- [x] Create one `plugin-runtime` package scaffold and migrate contracts + QuickJS core into it
 - [ ] Move worker wrapper/client transport and host adapter interfaces into `plugin-runtime` for non-UI embedding
 - [ ] Move Redux runtime policy/reducer/selector logic into `plugin-runtime` internal `redux-adapter` module
 - [ ] Refactor Playground into modular developer workbench UI shells (`catalog`, `workspace`, `inspector`)
